requirements.txt
asgiref==3.11.1
Django==6.0.2
docxtpl==0.20.2
Jinja2==3.1.6
lxml==6.0.2
MarkupSafe==3.0.3
python-docx==1.2.0
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
